+++ android/src/main/java/com/flurry/android/reactnative/FlurryModule.java
@@ -29,23 +29,33 @@ import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.ReadableMapKeySetIterator;
 import com.facebook.react.bridge.WritableMap;
+import com.facebook.react.modules.core.DeviceEventManagerModule;
 import com.facebook.react.uimanager.IllegalViewOperationException;
 import com.flurry.android.Constants;
 import com.flurry.android.FlurryAgent;
 import com.flurry.android.FlurryAgentListener;
+import com.flurry.android.marketing.FlurryMarketingModule;
+import com.flurry.android.marketing.FlurryMarketingOptions;
+import com.flurry.android.marketing.messaging.FlurryMessagingListener;
+import com.flurry.android.marketing.messaging.notification.FlurryMessage;
 
 import java.util.HashMap;
 import java.util.Map;
 import android.util.Log;
 
 public class FlurryModule extends ReactContextBaseJavaModule {
+    private static final String TAG = "FlurryModule";
 
     private static final String REACT_CLASS = "ReactNativeFlurry";
+    private static final String FLURRY_MESSAGING_EVENT = "FlurryMessagingEvent";
 
     private static final String ORIGIN_NAME = "react-native-flurry-sdk";
-    private static final String ORIGIN_VERSION = "2.1.0";
+    private static final String ORIGIN_VERSION = "3.0.0";
 
     private FlurryAgent.Builder mFlurryAgentBuilder;
+
+    private static ReactApplicationContext sReactApplicationContext = null;
+    private static boolean sEnableMessagingListener = false;
 
     @Override
     public String getName() {
@@ -54,6 +64,8 @@ public class FlurryModule extends ReactContextBaseJavaModule {
 
     public FlurryModule(ReactApplicationContext reactContext) {
         super(reactContext);
+
+        sReactApplicationContext = reactContext;
 
         // TODO remove this call after non-Builder APIs removed.
         initBuilder();
@@ -107,6 +119,11 @@ public class FlurryModule extends ReactContextBaseJavaModule {
     }
 
     @ReactMethod
+    public void withMessaging(boolean enableMessaging) {
+        Log.i(TAG, "To enable Flurry Push for Android, please duplicate Builder setup in your MainApplication.java.");
+    }
+
+    @ReactMethod
     public void setAge(int age) {
         FlurryAgent.setAge(age);
     }
@@ -144,7 +161,7 @@ public class FlurryModule extends ReactContextBaseJavaModule {
 
     @ReactMethod
     public void setIAPReportingEnabled(boolean enableIAP) {
-        Log.i("FlurryModule", "setIAPReportingEnabled is not supported on Android. Please use logPayment instead.");
+        Log.i(TAG, "setIAPReportingEnabled is not supported on Android. Please use logPayment instead.");
     }
 
     @ReactMethod
@@ -246,6 +263,16 @@ public class FlurryModule extends ReactContextBaseJavaModule {
         FlurryAgent.onPageView();
     }
 
+    @ReactMethod
+    public void enableMessagingListener(boolean enable) {
+        sEnableMessagingListener = enable;
+    }
+
+    @ReactMethod
+    public void willHandleMessage(boolean handled) {
+        RNFlurryMessagingListener.notifyCallbackReturn(handled);
+    }
+
     private static Map<String, String> toMap(final ReadableMap readableMap) {
         if (readableMap == null) {
             return null;
@@ -263,6 +290,218 @@ public class FlurryModule extends ReactContextBaseJavaModule {
         }
 
         return result;
+    }
+
+    /**
+     * Builder Pattern class for Flurry. Used by MainApplication to initialize Flurry Push for messaging.
+     */
+    public static class Builder {
+        private FlurryAgent.Builder mFlurryAgentBuilder;
+
+        public Builder() {
+            mFlurryAgentBuilder = new FlurryAgent.Builder();
+        }
+
+        /**
+         * True to enable or  false to disable the ability to catch all uncaught exceptions
+         * and have them reported back to Flurry.
+         *
+         * @param captureExceptions true to enable, false to disable.
+         * @return The Builder instance.
+         */
+        public Builder withCrashReporting(final boolean captureExceptions) {
+            mFlurryAgentBuilder.withCaptureUncaughtExceptions(captureExceptions);
+            return this;
+        }
+
+        /**
+         * Set the timeout for expiring a Flurry session.
+         *
+         * @param sessionMillis The time in milliseconds to set the session timeout to. Minimum value of 5000.
+         * @return The Builder instance.
+         */
+        public Builder withContinueSessionMillis(final long sessionMillis) {
+            mFlurryAgentBuilder.withContinueSessionMillis(sessionMillis);
+            return this;
+        }
+
+        /**
+         * True if this session should be added to total sessions/DAUs when applicationstate is inactive or background.
+         * Default is set to true.
+         *
+         * @param includeBackgroundSessionsInMetrics if background and inactive session should be counted toward dau
+         */
+        public Builder withIncludeBackgroundSessionsInMetrics(final boolean includeBackgroundSessionsInMetrics) {
+            mFlurryAgentBuilder.withIncludeBackgroundSessionsInMetrics(includeBackgroundSessionsInMetrics);
+            return this;
+        }
+
+        /**
+         * True to enable or false to disable the internal logging for the Flurry SDK.
+         *
+         * @param enableLog true to enable logging, false to disable it.
+         * @return The Builder instance.
+         */
+        public Builder withLogEnabled(final boolean enableLog) {
+            mFlurryAgentBuilder.withLogEnabled(enableLog);
+            return this;
+        }
+
+        /**
+         * Set the log level of the internal Flurry SDK logging.
+         *
+         * @param logLevel The level to set it to.
+         * @return The Builder instance.
+         */
+        public Builder withLogLevel(final int logLevel) {
+            mFlurryAgentBuilder.withLogLevel(logLevel);
+            return this;
+        }
+
+        /**
+         * Enable Flurry add-on Messaging.
+         *
+         * @param enableMessaging true to enable messaging,
+         *                        currently support only auto integration.
+         * @return The Builder instance.
+         */
+        public Builder withMessaging(final boolean enableMessaging) {
+            if (!enableMessaging) {
+                return this;
+            }
+
+            RNFlurryMessagingListener messagingListener = new RNFlurryMessagingListener();
+
+            FlurryMarketingOptions flurryMessagingOptions = new FlurryMarketingOptions.Builder()
+                    .setupMessagingWithAutoIntegration()
+                    .withFlurryMessagingListener(messagingListener)
+                    // Define yours if needed
+                    // .withDefaultNotificationChannelId(NOTIFICATION_CHANNEL_ID)
+                    // .withDefaultNotificationIconResourceId(R.mipmap.ic_launcher_round)
+                    // .withDefaultNotificationIconAccentColor(getResources().getColor(R.color.colorPrimary))
+                    .build();
+
+            FlurryMarketingModule marketingModule = new FlurryMarketingModule(flurryMessagingOptions);
+            mFlurryAgentBuilder.withModule(marketingModule);
+
+            return this;
+        }
+
+        public void build(@NonNull final Context context, @NonNull final String apiKey) {
+            mFlurryAgentBuilder.build(context, apiKey);
+        }
+    }
+
+    /**
+     * Wrapper Flurry Messaging listenet.
+     */
+    static class RNFlurryMessagingListener implements FlurryMessagingListener {
+        private volatile static boolean sCallbackReturnValue = false;
+        private volatile static boolean sIsCallbackReturn = false;
+
+        enum EventType {
+            NotificationReceived("NotificationReceived"),
+            NotificationClicked("NotificationClicked"),
+            NotificationCancelled("NotificationCancelled"),
+            TokenRefresh("TokenRefresh"),
+            NonFlurryNotificationReceived("NonFlurryNotificationReceived");
+
+            private final String name;
+
+            EventType(String name) {
+                this.name = name;
+            }
+
+            public String getName() {
+                return name;
+            }
+        }
+
+        @Override
+        public boolean onNotificationReceived(FlurryMessage flurryMessage) {
+            if (sEnableMessagingListener && (sReactApplicationContext != null)) {
+                return sendEvent(EventType.NotificationReceived, flurryMessage, true);
+            }
+            return false;
+        }
+
+        @Override
+        public boolean onNotificationClicked(FlurryMessage flurryMessage) {
+            if (sEnableMessagingListener && (sReactApplicationContext != null)) {
+                return sendEvent(EventType.NotificationClicked, flurryMessage, true);
+            }
+            return false;
+        }
+
+        @Override
+        public void onNotificationCancelled(FlurryMessage flurryMessage) {
+            if (sEnableMessagingListener && (sReactApplicationContext != null)) {
+                sendEvent(EventType.NotificationCancelled, flurryMessage, false);
+            }
+        }
+
+        @Override
+        public void onTokenRefresh(String token) {
+            if (sEnableMessagingListener && (sReactApplicationContext != null)) {
+                sendEvent(EventType.TokenRefresh, token);
+            }
+        }
+
+        @Override
+        public void onNonFlurryNotificationReceived(Object message) {
+            // no-op
+        }
+
+        private boolean sendEvent(EventType type, FlurryMessage flurryMessage, boolean waitReturn) {
+            WritableMap params = Arguments.createMap();
+            params.putString("Type", type.getName());
+            params.putString("Title", flurryMessage.getTitle());
+            params.putString("Body", flurryMessage.getBody());
+            params.putString("ClickAction", flurryMessage.getClickAction());
+
+            Map<String, String> appData = flurryMessage.getAppData();
+            WritableMap data = Arguments.createMap();
+            for (String key : appData.keySet()) {
+                data.putString(key, appData.get(key));
+            }
+            params.putMap("Data", data);
+
+            sCallbackReturnValue = false;
+            sIsCallbackReturn = !waitReturn;
+            sReactApplicationContext.getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
+                    .emit(FLURRY_MESSAGING_EVENT, params);
+            waitCallbackReturn();
+            return sCallbackReturnValue;
+        }
+
+        private void sendEvent(EventType type, String token) {
+            WritableMap params = Arguments.createMap();
+            params.putString("Type", type.getName());
+            params.putString("Token", token);
+
+            sReactApplicationContext.getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
+                    .emit(FLURRY_MESSAGING_EVENT, params);
+        }
+
+        private static void waitCallbackReturn() {
+            synchronized (sReactApplicationContext) {
+                if (!sIsCallbackReturn) {
+                    try {
+                        sReactApplicationContext.wait(300);
+                    } catch (InterruptedException e) {
+                        Log.e(TAG, "Interrupted Exception!", e);
+                    }
+                }
+            }
+        }
+
+        private static void notifyCallbackReturn(boolean returnValue) {
+            synchronized (sReactApplicationContext) {
+                sCallbackReturnValue = returnValue;
+                sIsCallbackReturn = true;
+                sReactApplicationContext.notifyAll();
+            }
+        }
     }
 
 }